finish bot

+++ course_project/db/migration/000001_init_schema.up.sql
@@ -1,7 +1,7 @@
 create table applications (
     id bigserial primary key,
     ticker text not null,
-    cost integer not null,
+    cost float not null,
     size integer not null,
     created_at timestamp default now(),
     type text not null
